Добавил проверку на наличие lock-файла перед записью в него

+++ CardsController.php
@@ -4,6 +4,10 @@ class CardsController
     public static function lockCard($number, $student)
     {
         $lockFilePath = Cards::CARDS_DIR . DIRECTORY_SEPARATOR . $number . '.' . Cards::LOCK_CARD_EXT;
-        return file_put_contents($lockFilePath, $student);
+        if (!file_exists($lockFilePath)) {
+            return file_put_contents($lockFilePath, $student);
+        } else {
+            return false;
+        }
     }
 }
